requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
django-jsonform==2.19.1
djangorestframework==3.14.0
install==1.3.5
Markdown==3.5
Pillow==10.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
